Invoke hook subprocess as script instead of module
Module-based invocation works well when the subprocess interpreter has
the same Python PATH and modules as the main process, but this isn't
always the case (i.e. pytest).

It's safer to import the module we want to call in the main process and
directly invoke that script in the subprocess.

One behavioral difference between the module and script is that scripts
resolve import starting with the script's directory. To avoid
collisions, this change moves the scripts into a new subdirectory.